fix: Issue #694 (#711) Add tabs reconsile when merging profile.
* fix: Issue #694
Add tabs reconsile when merging profile.
