+++ src/impl/source/profiles/profileSync.ts
@@ -109,27 +109,4 @@ export default class ProfileSync extends ProfileActions {
         //Retun final status
         return profileStatus;
     }
-
-    private async reconcileTabs(profileObj: Profile): Promise<void> {
-        let tabRetriever = new MetadataRetriever(this.org.getConnection(), registry.types.customtab.name);
-
-        if (profileObj.tabVisibilities !== undefined) {
-            if (!Array.isArray(profileObj.tabVisibilities)) {
-                profileObj.tabVisibilities = [profileObj.tabVisibilities];
-            }
-            let validArray = [];
-            for (let i = 0; i < profileObj.tabVisibilities.length; i++) {
-                let cmpObj = profileObj.tabVisibilities[i];
-                let exist = await tabRetriever.isComponentExistsInProjectDirectoryOrInOrg(cmpObj.tab);
-                if (exist) {
-                    validArray.push(cmpObj);
-                }
-            }
-            SFPLogger.log(
-                `Tab Visibilities reduced from ${profileObj.tabVisibilities.length}  to  ${validArray.length}`,
-                LoggerLevel.DEBUG
-            );
-            profileObj.tabVisibilities = validArray;
-        }
-    }
 }
